[CHORE] Async 적용을 위한 클래스 분리
